Data.Boolean.Overload

Description

Author : Alex Horsman (aninhumer) Maintainer : conal@conal.net Stability : experimental

Definitions of Prelude function names in terms of their corresponding Data.Boolean generalised implementation. This can then be used as part of a partial or complete Prelude replacement.

Also exports ifThenElse for use with RebindableSyntax.

Documentation

module Data.Boolean

(&&) :: Boolean a => a -> a -> a infixr 3

(||) :: Boolean a => a -> a -> a infixr 2

not :: Boolean a => a -> a

ifThenElse :: IfB a => BooleanOf a -> a -> a -> a

(==) :: EqB a => a -> a -> BooleanOf a infix 4

(/=) :: EqB a => a -> a -> BooleanOf a infix 4

(<) :: OrdB a => a -> a -> BooleanOf a infix 4

(>) :: OrdB a => a -> a -> BooleanOf a infix 4

(<=) :: OrdB a => a -> a -> BooleanOf a infix 4

(>=) :: OrdB a => a -> a -> BooleanOf a infix 4

min :: (IfB a, OrdB a) => a -> a -> a

max :: (IfB a, OrdB a) => a -> a -> a